+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".Pages.Section_Forraje"
+            android:name=".Pages.Manejo_Section_Forraje"
             android:label=""
             android:exported="false"/>
         <activity
@@ -22,15 +22,15 @@
             android:label=""
             android:screenOrientation="fullUser"/>
         <activity
-            android:name=".Pages.RendimientoCalculo"
+            android:name=".Pages.Manejo_RendimientoCalculo"
             android:label=""
             android:exported="false"/>
         <activity
-            android:name=".Pages.PesoCalculo"
+            android:name=".Pages.Manejo_PesoCalculo"
             android:label=""
             android:exported="false" />
         <activity
-            android:name=".Pages.Manejo"
+            android:name=".Pages.Manejo_Inicio"
             android:label=""
             android:exported="false" />
         <activity
