+++ app/src/main/java/com/example/app/Pages/IndexPage.kt
@@ -3,7 +3,6 @@ package com.example.app.Pages
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.FrameLayout
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.example.app.R
 
@@ -11,7 +10,7 @@ class IndexPage : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_index_page)
+        setContentView(R.layout.inicio)
 
         val frameLayout: ConstraintLayout = findViewById(R.id.frame_layout)
         frameLayout.setOnClickListener{
