+++ app/src/main/java/com/example/app/Pages/Reproduccion_Inicio.kt
@@ -1,0 +1,6 @@
+package com.example.app.Pages
+
+import androidx.appcompat.app.AppCompatActivity
+
+class Reproduccion_Inicio : AppCompatActivity(){
+}
