+++ app/src/main/res/layout/activity_manejo.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FDF9F9"
-    tools:context=".Pages.Manejo">
+    tools:context=".Pages.Manejo_Inicio">
 
     <LinearLayout
         android:layout_width="match_parent"
