+++ app/src/main/res/layout/manejo_peso_calculo.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FDF9F9"
-    tools:context=".Pages.PesoCalculo">
+    tools:context=".Pages.Manejo_PesoCalculo">
 
     <LinearLayout
         android:layout_width="match_parent"
